currently I found that the approx_math sine is even better than the dsp sine, with a slightly higher computation time but better accuracy. I think the major problem of the dsp is that it takes and returns q15/q31, the conversion between it and integer is time-consuming. I will take a look at CORDIC and do more testing before the final selection.

+++ STM32F405VG/core/dsp_math.h
@@ -27,7 +27,7 @@
 #include <stdbool.h>
 #include <string.h>
 
-int16_t dsp_cos(int16_t angle);
-int16_t dsp_sin(int16_t angle);
+int16_t dsp_cos(int32_t angle);
+int16_t dsp_sin(int32_t angle);
 
 #endif
